Move parameter pack to its own header
